+++ module/data/item-ranged-weapon.mjs
@@ -8,9 +8,9 @@ export default class DieseldrachenRangedWeapon extends DieseldrachenItemBase {
 
     schema.weight = new fields.NumberField({ required: true, nullable: false, integer: true, initial: 1, min: 0, max: 100 });
     schema.weaponType = new fields.StringField({ required: true, blank: true, initial: "pistol" });
-    schema.precision_near = new fields.NumberField({ required: true, nullable: false, integer: true, initial: 0, min: 0, max: 100 });
-    schema.precision_close = new fields.NumberField({ required: true, nullable: false, integer: true, initial: 0, min: 0, max: 100 });
-    schema.precision_far = new fields.NumberField({ required: true, nullable: false, integer: true, initial: 0, min: 0, max: 100 });
+    schema.precision_near = new fields.NumberField({ required: true, nullable: false, integer: true, initial: 4, min: 0, max: 100 });
+    schema.precision_close = new fields.NumberField({ required: true, nullable: false, integer: true, initial: 4, min: 0, max: 100 });
+    schema.precision_far = new fields.NumberField({ required: true, nullable: false, integer: true, initial: 4, min: 0, max: 100 });
 
     schema.damage_near = new fields.NumberField({ required: true, nullable: false, integer: true, initial: 0, min: 0, max: 100 });
     schema.damage_close = new fields.NumberField({ required: true, nullable: false, integer: true, initial: 0, min: 0, max: 100 });